backend/requirements.txt
fastapi
uvicorn
langgraph
openai
python-dotenv
requests
pydantic